fix: Axis Styles feat: Add description refactor: move files refactor: extract DOM function feat: Based on type, add custom styles to iframe
