+++ index.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset="UTF-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+    <meta name="description" content="Dual map - synchronize views of different map types side by side. Compare data from different sources."/>
     <title>Dual Map</title>
     <script type="module" src="/src/main.ts"></script>
     <link rel="icon" type="image/svg+xml"
