--- src/dom.ts
@@ -1,2 +1,0 @@
-export const $: typeof document.querySelector = document.querySelector.bind(document);
-export const $$: typeof document.querySelectorAll = document.querySelectorAll.bind(document);
